Clean up inconsistencies between models and DB tables
Changed a few property names in models to make them
consistent with the names in DB tables and made several
tweaks to the data and stored procedures.

+++ DnDCharacterBuilderDB/dbo/Tables/RacialFeatures.sql
@@ -5,6 +5,6 @@
     [RaceId] TINYINT NULL, 
     [Description] VARCHAR(5000) NULL, 
     [ParentFeatureName] VARCHAR(80) NULL, 
-    [SubfeatureSelections] TINYINT NULL, 
+    [SubfeatureSelections] SMALLINT NULL, 
     [Tags] VARCHAR(200) NULL
 )
